+++ code/src/cljs/sixsq/nuvla/ui/main/views.cljs
@@ -37,6 +37,7 @@
     [sixsq.nuvla.ui.utils.semantic-ui-extensions :as uix]
     [sixsq.nuvla.ui.utils.time :as time]
     [sixsq.nuvla.ui.utils.ui-callback :as ui-callback]
+    [sixsq.nuvla.ui.profile.subs :as profile-subs]
     [sixsq.nuvla.ui.welcome.views]))
 
 
@@ -232,7 +233,8 @@
         iframe?          (subscribe [::subs/iframe?])
         is-small-device? (subscribe [::subs/is-small-device?])
         resource-path    (subscribe [::subs/nav-path])
-        session-loading? (subscribe [::session-subs/session-loading?])]
+        session-loading? (subscribe [::session-subs/session-loading?])
+        subs-canceled?   (subscribe [::profile-subs/subscription-canceled?])]
     (if (and @cep (not @session-loading?))
       [:div {:id "nuvla-ui-main"}
        (case (first @resource-path)
@@ -254,6 +256,14 @@
                        :on-click #(dispatch [::events/close-sidebar])}]
            [header]
            [:div {:ref main-components/ref}
+            (when @subs-canceled?
+              [uix/Message {:icon    "warning"
+                            :header  "Subscription is canceled"
+                            :content [:span
+                                      "To reactivate your subscription "
+                                      [:a {:href "profile"} "go to the profile"]
+                                      " and make sure to have a payment method."]
+                            :type    :error}])
             [contents]
             [ignore-changes-modal]
             [subscription-required-modal]